+++ internal/service/base/resource_notification_template_content_test.go
@@ -67,6 +67,76 @@ func testAccCheckNotificationTemplateContentDestroy(s *terraform.State) error {
 	return nil
 }
 
+func testAccGetNotificationTemplateContentIDs(resourceName string, environmentID, templateName, resourceID *string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return fmt.Errorf("Resource Not found: %s", resourceName)
+		}
+
+		*resourceID = rs.Primary.ID
+		*templateName = rs.Primary.Attributes["template_name"]
+		*environmentID = rs.Primary.Attributes["environment_id"]
+
+		return nil
+	}
+}
+
+func TestAccNotificationTemplateContent_RemovalDrift(t *testing.T) {
+	t.Parallel()
+
+	resourceName := acctest.ResourceNameGen()
+	resourceFullName := fmt.Sprintf("pingone_notification_template_content.%s", resourceName)
+
+	environmentName := acctest.ResourceNameGenEnvironment()
+
+	licenseID := os.Getenv("PINGONE_LICENSE_ID")
+
+	name := "strong_authentication"
+	locale := "en-GB"
+
+	var resourceID, templateName, environmentID string
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheckEnvironment(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckNotificationTemplateContentDestroy,
+		ErrorCheck:               acctest.ErrorCheck(t),
+		Steps: []resource.TestStep{
+			// Configure
+			{
+				Config: testAccNotificationTemplateContentConfig_NewLocale_Minimal(environmentName, licenseID, resourceName, name, locale),
+				Check:  testAccGetNotificationTemplateContentIDs(resourceFullName, &environmentID, &templateName, &resourceID),
+			},
+			// Replan after removal preconfig
+			{
+				PreConfig: func() {
+					var ctx = context.Background()
+					p1Client, err := acctest.TestClient(ctx)
+
+					if err != nil {
+						t.Fatalf("Failed to get API client: %v", err)
+					}
+
+					apiClient := p1Client.API.ManagementAPIClient
+
+					if environmentID == "" || templateName == "" || resourceID == "" {
+						t.Fatalf("One of environment ID, template name or resource ID cannot be determined. Environment ID: %s, Template Name: %s, Resource ID: %s", environmentID, templateName, resourceID)
+					}
+
+					_, err = apiClient.NotificationsTemplatesApi.DeleteContent(ctx, environmentID, templateName, resourceID).Execute()
+					if err != nil {
+						t.Fatalf("Failed to delete notification template contents: %v", err)
+					}
+				},
+				RefreshState:       true,
+				ExpectNonEmptyPlan: true,
+			},
+		},
+	})
+}
+
 func TestAccNotificationTemplateContent_OverrideDefaultLocale(t *testing.T) {
 	t.Parallel()