+++ internal/service/sso/resource_resource_attribute_test.go
@@ -67,6 +67,71 @@ func testAccCheckResourceAttributeDestroy(s *terraform.State) error {
 	return nil
 }
 
+func testAccGetResourceAttributeIDs(resourceName string, environmentID, oidcResourceID, resourceID *string) resource.TestCheckFunc {
+	return func(s *terraform.State) error {
+
+		rs, ok := s.RootModule().Resources[resourceName]
+		if !ok {
+			return fmt.Errorf("Resource Not found: %s", resourceName)
+		}
+
+		*resourceID = rs.Primary.ID
+		*oidcResourceID = rs.Primary.Attributes["resource_id"]
+		*environmentID = rs.Primary.Attributes["environment_id"]
+
+		return nil
+	}
+}
+
+func TestAccResourceAttribute_RemovalDrift(t *testing.T) {
+	t.Parallel()
+
+	resourceName := acctest.ResourceNameGen()
+	resourceFullName := fmt.Sprintf("pingone_resource_attribute.%s", resourceName)
+
+	name := resourceName
+
+	var resourceID, oidcResourceID, environmentID string
+
+	resource.Test(t, resource.TestCase{
+		PreCheck:                 func() { acctest.PreCheckEnvironment(t) },
+		ProtoV6ProviderFactories: acctest.ProtoV6ProviderFactories,
+		CheckDestroy:             testAccCheckResourceAttributeDestroy,
+		ErrorCheck:               acctest.ErrorCheck(t),
+		Steps: []resource.TestStep{
+			// Configure
+			{
+				Config: testAccResourceAttributeConfig_OIDC_Custom_Minimal(resourceName, name),
+				Check:  testAccGetResourceAttributeIDs(resourceFullName, &environmentID, &oidcResourceID, &resourceID),
+			},
+			// Replan after removal preconfig
+			{
+				PreConfig: func() {
+					var ctx = context.Background()
+					p1Client, err := acctest.TestClient(ctx)
+
+					if err != nil {
+						t.Fatalf("Failed to get API client: %v", err)
+					}
+
+					apiClient := p1Client.API.ManagementAPIClient
+
+					if environmentID == "" || oidcResourceID == "" || resourceID == "" {
+						t.Fatalf("One of environment ID or resource ID cannot be determined. Environment ID: %s,OIDC Resource ID: %s, Resource ID: %s", environmentID, oidcResourceID, resourceID)
+					}
+
+					_, err = apiClient.ResourceAttributesApi.DeleteResourceAttribute(ctx, environmentID, oidcResourceID, resourceID).Execute()
+					if err != nil {
+						t.Fatalf("Failed to delete resource attribute: %v", err)
+					}
+				},
+				RefreshState:       true,
+				ExpectNonEmptyPlan: true,
+			},
+		},
+	})
+}
+
 func TestAccResourceAttribute_OIDC_Custom(t *testing.T) {
 	t.Parallel()